Add minimalist C shared library crate
This shared object provides a C API/ABI for a dynamically linked library.

It is deliberately minimal.

The caller handles only serialized forms of the underlying cryptographic
objects, and is responsible for all memory allocation/destruction.

A more sophisticated library interface seems possible, but this one is
clean and can be easily bound to other programming languages that already
have the ability to wrap a C FFI.